Large update handling FF and logic separatly.
First, gates are classified either as logic or flip-flop (FF, if they have a clock).

Then, all the logic is grouped in clusters in which:
- All gates are daisy chained to constitute the critical path (CP).
- One input is reserved for an incoming FF.
- The last gate in the chain is connected to n (fanout) FFs.
- All the remaining input port on the gates are interconnected with the ouputs of the other gates in the cluster, at random.

Create a fixed amount of I/O based on Rent's Rule: T = t * g^p

Those I/O are connected to available FF.

All remaining FF are interconnected together at random. This might be a bit problematic if we need something more elaborate. This might create some sort of a macro bloc in the netlist that might be a problem for the placement.